motion: replace lineno as motion id with opaque serial counter + side table
- IDL (motctl/motstat/tp): id fields i64 → i32
- Regenerate all 9 generated GMI files
- C structs: int64_t → int for id in emcmot_command_t, emcmot_status_t,
  emcmot_internal_t, TC_STRUCT, TP_STRUCT; MOTION_INVALID_ID INT64_MIN → INT_MIN;
  remove now-unneeded <stdint.h> includes
- C callsites: handler signatures int64_t → int32_t; PRId64 → %d throughout
  tp.c, tc.c, command.c, motctl_handlers.c, motion-logger.c
- Error messages: 'move on line %d' / 'at line %d' → '(segment %d)' since id
  is now an opaque serial, not a G-code line number
- HAL pin: rename motion.program-line → motion.segment-id (mot_hal_types.h,
  motion.c, control.c)

Go side:
- MotionController interface: all id int64 → int32
- Canon: add nextSerial int32 (starts at 1; 0 reserved as idle sentinel);
  allocSerial() increments counter, calls interp.FileName() for correct file
  name (handles o-word subroutine files), and registers in side table
- All motion cmd ID fields (LinearMoveCmd, CircularMoveCmd, RigidTapCmd,
  ProbeCmd): int64 → int32; all ID: int64(lineno) → c.allocSerial(lineno)
- Task: add motionMap map[int32]motionInfo side table with registerMotion,
  lookupMotionLine, pruneMotionMap helpers; initialise in NewTask
- stat.go: MotionLine = lookupMotionLine(ms.Id) instead of ms.Id; prune map
- halui.go: add halui.program-line (int32 out) pin; write from side table
- sequencer.go: currentLine = lookupMotionLine(id) instead of raw serial
- Tests: mock signatures int64 → int32

UI:
- axis.py: set_current_line(stat.motion_line) — was motion_id or motion_line
- touchy/emc_interface.py: highlight_line and status 'id' display use
  motion_line instead of motion_id

+++ src/emc/motion/motctl_handlers.c
@@ -90,7 +90,7 @@ static inline EmcPose pose_from_motctl(const motctl_pose_t *p)
 
 static int32_t h_set_line(void *ctx, const motctl_pose_t *pos,
     double vel, double ini_maxvel, double acc,
-    int32_t motion_type, int64_t id, double feed_upm,
+    int32_t motion_type, int32_t id, double feed_upm,
     int32_t indexer_jnum)
 {
     CTX; emcmot_command_t cmd;
@@ -109,7 +109,7 @@ static int32_t h_set_line(void *ctx, const motctl_pose_t *pos,
 static int32_t h_set_circle(void *ctx, const motctl_pose_t *pos,
     const motctl_cartesian_t *center, const motctl_cartesian_t *normal,
     int32_t turn, double vel, double ini_maxvel, double acc,
-    int32_t motion_type, int64_t id, double feed_upm)
+    int32_t motion_type, int32_t id, double feed_upm)
 {
     CTX; emcmot_command_t cmd;
     cmd_init(&cmd, EMCMOT_SET_CIRCLE);
@@ -132,7 +132,7 @@ static int32_t h_set_circle(void *ctx, const motctl_pose_t *pos,
 
 static int32_t h_probe(void *ctx, const motctl_pose_t *pos,
     double vel, double ini_maxvel, double acc,
-    int32_t motion_type, uint8_t probe_type, int64_t id, double feed_upm)
+    int32_t motion_type, uint8_t probe_type, int32_t id, double feed_upm)
 {
     CTX; emcmot_command_t cmd;
     cmd_init(&cmd, EMCMOT_PROBE);
@@ -149,7 +149,7 @@ static int32_t h_probe(void *ctx, const motctl_pose_t *pos,
 
 static int32_t h_rigid_tap(void *ctx, const motctl_pose_t *pos,
     double vel, double ini_maxvel, double acc,
-    double scale, int64_t id, double feed_upm)
+    double scale, int32_t id, double feed_upm)
 {
     CTX; emcmot_command_t cmd;
     cmd_init(&cmd, EMCMOT_RIGID_TAP);
@@ -242,7 +242,7 @@ static int32_t h_resume(void *ctx)
     return send_command(mc, &cmd);
 }
 
-static int32_t h_step(void *ctx, int64_t id)
+static int32_t h_step(void *ctx, int32_t id)
 {
     CTX; emcmot_command_t cmd;
     cmd_init(&cmd, EMCMOT_STEP);
