+++ src/emc/tp/tc_types.h
@@ -17,7 +17,6 @@
 #include "posemath.h"
 #include "emcpos.h"
 #include "emcmotcfg.h"
-#include <stdint.h>
 
 #define BLEND_DIST_FRACTION 0.5
 /* values for endFlag */
@@ -138,7 +137,7 @@ typedef struct {
     double maxaccel;        // accel calc'd by task
     double acc_ratio_tan;// ratio between normal and tangential accel
     
-    int64_t id;                 // segment's serial number
+    int id;                 // segment's serial number
     double feed_upm;            // programmed feed rate in G-code units per minute
 
     union {                 // describes the segment's start and end positions
